ci: only run msbuild if needed
Each build may only take about 2 minutes, but we (already) build 6 variants (ctrlr X ros2_version) and it'd be wasteful to do that if not actually needed.

This change adds a filter job which sole purpose is to check whether any files in the 'src' directory have changed, and/or the VS solution itself. Only in those cases will the msbuild job itself be run.

Changes to other files and/or directories (such as documentation) will not trigger a full build of the VS proj.